Moved PropulsionDemo to seperate folder
