+++ src/aircraft_data_hierarchy/performanceLib/propulsion/propulsion_performance_builder.py
@@ -387,346 +387,3 @@ class pyCycleBuilder(PropulsionPerformanceBuilder):
 class NPSSBuilder(PropulsionPerformanceBuilder):
     def __init__(self, ADHInstance):
         raise Exception("NPSS Builder not implemented!")
-
-
-# Temporary demo solution
-def HBTFprep(prob, ADHInstance):
-    if utils.lenient_isinstance(ADHInstance.cycle, MultiPointCycle):
-        cycle = ADHInstance.cycle.design_point
-    else:
-        cycle = ADHInstance.cycle
-    behavior = ADHInstance.behavior
-    # Set PR and eff
-
-    prob.set_val("DESIGN.fan.PR", cycle.elements[1].pr_des)
-    prob.set_val("DESIGN.fan.eff", cycle.elements[1].eff_des)
-
-    prob.set_val("DESIGN.lpc.PR", cycle.elements[4].pr_des)
-    prob.set_val("DESIGN.lpc.eff", cycle.elements[4].eff_des)
-
-    prob.set_val("DESIGN.hpc.PR", cycle.elements[6].pr_des)
-    prob.set_val("DESIGN.hpc.eff", cycle.elements[6].eff_des)
-
-    prob.set_val("DESIGN.hpt.eff", cycle.elements[9].eff_des)
-    prob.set_val("DESIGN.lpt.eff", cycle.elements[11].eff_des)
-
-    # print(cycle.elements[1].pr_des)
-    # print(cycle.elements[1].eff_des)
-    # print(cycle.elements[4].pr_des)
-    # print(cycle.elements[4].eff_des)
-    # print(cycle.elements[6].pr_des)
-    # print(cycle.elements[6].eff_des)
-    # print(cycle.elements[9].eff_des)
-    # print(cycle.elements[11].eff_des)
-
-    prob.set_val("DESIGN.fc.alt", behavior.flight_conditions_design.alt[0], units="ft")
-    prob.set_val("DESIGN.fc.MN", behavior.flight_conditions_design.mn[0])
-
-    # print(behavior.flight_conditions_design.alt[0])
-    # print(behavior.flight_conditions_design.mn[0])
-
-    prob.set_val("DESIGN.T4_MAX", 2857, units="degR")
-    prob.set_val("DESIGN.Fn_DES", 5900.0, units="lbf")  # TODO
-
-    prob.set_val("OD_full_pwr.T4_MAX", 2857, units="degR")  # TODO
-    prob.set_val("OD_part_pwr.PC", 0.8)
-
-    # Set initial guesses for balances
-    prob["DESIGN.balance.FAR"] = 0.025
-    prob["DESIGN.balance.W"] = 100.0
-    prob["DESIGN.balance.lpt_PR"] = 4.0
-    prob["DESIGN.balance.hpt_PR"] = 3.0
-    prob["DESIGN.fc.balance.Pt"] = 5.2
-    prob["DESIGN.fc.balance.Tt"] = 440.0
-
-    for pt in ADHInstance.cycle.od_points:
-
-        print(pt.name)
-
-        # initial guesses
-        prob[pt.name + ".balance.FAR"] = 0.02467
-        prob[pt.name + ".balance.W"] = 300
-        prob[pt.name + ".balance.BPR"] = 5.105
-        prob[pt.name + ".balance.lp_Nmech"] = 5000
-        prob[pt.name + ".balance.hp_Nmech"] = 15000
-        prob[pt.name + ".hpt.PR"] = 3.0
-        prob[pt.name + ".lpt.PR"] = 4.0
-        prob[pt.name + ".fan.map.RlineMap"] = 2.0
-        prob[pt.name + ".lpc.map.RlineMap"] = 2.0
-        prob[pt.name + ".hpc.map.RlineMap"] = 2.0
-
-    ODpoints = ADHInstance.cycle.od_points
-    machs = ODpoints[0].flight_conditions_od.mn
-    alts = ODpoints[0].flight_conditions_od.alt
-
-    flight_env = list(zip(machs, alts))
-
-    print(flight_env)
-
-    return prob, flight_env
-
-
-if __name__ == "__main__":
-
-    # Set-up the ADH Propulsion instance for demo purposes
-
-    metadata = Metadata(key="example_key", value="example_value")
-
-    fc = FlightConditions(name="fc", mn=[0.8], alt=[35000])
-    inlet = Inlet(name="inlet")
-    fan = Compressor(name="fan", map_data="FanMap", bleed_names=[], map_extrap=True)
-    splitter = Splitter(name="splitter")
-    duct4 = Duct(name="duct4")
-    lpc = Compressor(name="lpc", map_data="LPCMap", bleed_names=[], map_extrap=True)
-    duct6 = Duct(name="duct6")
-    hpc = Compressor(name="hpc", map_data="HPCMap", bleed_names=["cool1", "cool2", "cust"], map_extrap=True)
-    bld3 = Bleed(name="bld3", bleed_names=["cool3", "cool4"])
-    burner = Combustor(name="burner", fuel_type="Jet-A(g)")
-    hpt = Turbine(name="hpt", map_data="HPTMap", bleed_names=["cool3", "cool4"], map_extrap=True)
-    duct11 = Duct(name="duct11")
-    lpt = Turbine(name="lpt", map_data="LPTMap", bleed_names=["cool1", "cool2"], map_extrap=True)
-    duct13 = Duct(name="duct13")
-    core_nozz = Nozzle(name="core_nozz", nozz_type="CV", loss_coef="Cv")
-    byp_bld = Bleed(name="byp_bld", bleed_names=["bypBld"])
-    duct15 = Duct(name="duct15")
-    byp_nozz = Nozzle(name="byp_nozz", nozz_type="CV", loss_coef="Cv")
-    lp_shaft = Shaft(name="lp_shaft", nmech_type="LP", num_ports=3)
-    hp_shaft = Shaft(name="hp_shaft", nmech_type="HP", num_ports=2)
-
-    perf = Performance(
-        name="perf",
-        pt2_source="inlet",
-        pt3_source="hpc",
-        wfuel_0_source="burner",
-        ram_drag_source="inlet",
-        fg_0_source="core_nozz",
-        fg_1_source="byp_nozz",
-    )
-
-    # Set data
-
-    # Inlet
-    inlet.mn = 0.751
-    inlet.ram_recovery = 0.999
-
-    # Fan
-    fan.mn = 0.4578
-
-    # Comps
-    lpc.mn = 0.3059
-    hpc.mn = 0.2442
-    hpc.frac_W = [0.050708, 0.020274, 0.0445]
-    hpc.frac_P = [0.5, 0.55, 0.5]
-    hpc.frac_work = [0.5, 0.5, 0.5]
-
-    # Splitter
-    splitter.bpr = 5.105
-    splitter.mn1 = 0.3104
-    splitter.mn2 = 0.4518
-
-    # Ducts
-    duct4.mn = 0.3121
-    duct6.mn = 0.3563
-    duct11.mn = 0.3063
-    duct13.mn = 0.4463
-    duct15.mn = 0.4589
-
-    duct4.dPqP = 0.0048
-    duct6.dPqP = 0.0101
-    duct11.dPqP = 0.0051
-    duct13.dPqP = 0.0107
-    duct15.dPqP = 0.0149
-
-    # Bleeds
-    bld3.mn = 0.300
-    byp_bld.mn = 0.4489
-
-    bld3.frac_W = [0.067214, 0.101256]
-    byp_bld.frac_W = [0.005]
-
-    # Combs
-    burner.mn = 0.1025
-    burner.dp_qp = 0.0540
-
-    # Turbs
-
-    lpt.mn = 0.4127
-    hpt.mn = 0.3650
-
-    hpt.frac_P = [1.0, 0.0]
-    lpt.frac_P = [1.0, 0.0]
-
-    # Nozzle
-
-    core_nozz.cv = 0.9933
-    byp_nozz.cv = 0.9939
-
-    # Set PR and eff
-    fan.pr_des = 1.685
-    fan.eff_des = 0.8948
-
-    lpc.pr_des = 1.935
-    lpc.eff_des = 0.9243
-
-    hpc.pr_des = 9.369
-    hpc.eff_des = 0.8707
-
-    hpt.eff_des = 0.8888
-    lpt.eff_des = 0.8996
-
-    cycle = PropulsionCycle(
-        name="Cycle",
-        elements=[
-            inlet,
-            fan,
-            splitter,
-            duct4,
-            lpc,
-            duct6,
-            hpc,
-            bld3,
-            burner,
-            hpt,
-            duct11,
-            lpt,
-            duct13,
-            core_nozz,
-            byp_bld,
-            duct15,
-            byp_nozz,
-            lp_shaft,
-            hp_shaft,
-        ],
-        global_connections=["fan,lp_shaft", "lpc,lp_shaft", "lpt,lp_shaft", "hpc,hp_shaft", "hpt,hp_shaft"],
-        flow_connections=[
-            ["fc", "inlet"],
-            ["inlet", "fan"],
-            ["fan", "splitter"],
-            ["splitter", "duct4", "1"],
-            ["duct4", "lpc"],
-            ["lpc", "duct6"],
-            ["duct6", "hpc"],
-            ["hpc", "bld3"],
-            ["bld3", "burner"],
-            ["burner", "hpt"],
-            ["hpt", "duct11"],
-            ["duct11", "lpt"],
-            ["lpt", "duct13"],
-            ["duct13", "core_nozz"],
-            ["splitter", "byp_bld", "2"],
-            ["byp_bld", "duct15"],
-            ["duct15", "byp_nozz"],
-        ],
-    )
-
-    cyclePerformance = PropulsionCyclePerformance(
-        name="CyclePerformance",
-        thermo_method="CEA",
-        throttle_mode="T4",
-    )
-
-    cycleBehavior = PropulsionCycleBehavior(
-        name="CycleBehavior",
-        design=True,
-        flight_conditions_design=fc,
-        performance_components=[perf],
-    )
-
-    file_header = FileHeader(name="engineDeck", author=[Author(name="Safa Bakhshi")])
-
-    ODpoints = []
-    fc2 = FlightConditions(
-        name="od_full_pwr_fc",
-        mn=[0.8, 0.7, 0.4, 0.6, 0.8, 0.6, 0.4, 0.2, 0.001, 0.001, 0.2, 0.4, 0.6, 0.6, 0.4, 0.2, 0.001],
-        alt=[35000, 35000, 20000, 20000, 20000, 10000, 10000, 10000, 10000, 1000, 1000, 1000, 1000, 0, 0, 0, 0],
-        d_ts=0.0,
-    )
-    fc3 = FlightConditions(
-        name="od_part_pwr_fc",
-        mn=[0.8, 0.7, 0.4, 0.8, 0.6, 0.4, 0.6, 0.4, 0.2],
-        alt=[35000, 35000, 35000, 10000, 10000, 10000, 0, 0, 0],
-        d_ts=0.0,
-    )
-
-    od1 = OffDesignPoint(
-        name="OD_full_pwr",
-        flight_conditions_od=fc2,
-        throttle_mode="T4",
-    )
-
-    od2 = OffDesignPoint(
-        name="OD_part_pwr",
-        flight_conditions_od=fc3,
-        PC=[1.0, 0.8],
-        throttle_mode="percent_thrust",
-    )
-
-    multipoint = MultiPointCycle(
-        design_point=cycle,
-        od_points=[od1, od2],
-    )
-
-    ADHInstance = Propulsion(
-        name="Engine",
-        description="Main engine component",
-        subcomponents=[],
-        metadata=metadata,
-        cycle=multipoint,
-        performance=cyclePerformance,
-        behavior=cycleBehavior,
-    )
-
-    pycTest = pyCycleBuilder(ADHInstance)
-    pycTest.getInput()
-
-    prob = om.Problem()
-    prob.model = pycTest.getOutput()
-
-    prob.setup()
-
-    # USER SCRIPT FOR RUNNING ANALYSIS BELOW THIS LINE
-    # -----------------------------------------------
-    prob, flight_env = HBTFprep(prob, ADHInstance)
-    om.n2(prob, show_browser=False)
-    prob.set_solver_print(level=-1)
-    prob.set_solver_print(level=2, depth=1)
-
-    viewer_file = open("hbtf_view.out", "w")
-    first_pass = True
-    for MN, alt in flight_env:
-
-        # NOTE: You never change the MN,alt for the
-        # design point because that is a fixed reference condition.
-
-        print("***" * 10)
-        print(f"* MN: {MN}, alt: {alt}")
-        print("***" * 10)
-        prob["OD_full_pwr.fc.MN"] = MN
-        prob["OD_full_pwr.fc.alt"] = alt
-
-        prob["OD_part_pwr.fc.MN"] = MN
-        prob["OD_part_pwr.fc.alt"] = alt
-
-        for PC in ADHInstance.cycle.od_points[1].PC:
-            print(f"## PC = {PC}")
-            prob["OD_part_pwr.PC"] = PC
-            prob.run_model()
-
-            if first_pass:
-                # viewer(prob, "DESIGN", file=viewer_file)
-                first_pass = False
-            # viewer(prob, "OD_part_pwr", file=viewer_file)
-
-        # run throttle back up to full power
-        for PC in [1, 0.85]:
-            prob["OD_part_pwr.PC"] = PC
-            prob.run_model()
-
-    outputs = prob.model.list_outputs(
-        out_stream=None, residuals_tol=1e-2, implicit=True, explicit=False, residuals=True
-    )
-
-    from pprint import pprint
-
-    with open("output.txt", "w") as f:
-        pprint(outputs, stream=f)
